# docs/upstream-api/wg-guard-openapi.json
{
  "openapi": "3.2.1",
  "info": {
    "title": "WG-Guard API",
...
    "version": "1.0.0"
  },
  "servers": [{ "url": "/" }],
  "security": [{ "BearerAuth": [] }],
...
  "paths": {
...
    "/api/v1/webhooks": {
      "get": {
        "tags": ["webhooks"], "summary": "List webhook endpoints (with delivery stats)", "security": [{ "BearerAuth": ["webhooks.read"] }],
        "responses": { "200": { "description": "Endpoints", "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/WebhookEndpoint" } } } } } } } }
      },
      "post": {
        "tags": ["webhooks"], "summary": "Create a webhook endpoint", "description": "An empty `secret` generates one and returns it exactly once. Signing: `X-WG-Signature: t=<ts>,v1=<hex hmac-sha256(secret, \"<ts>.<body>\")>` with a ~5-minute replay window.", "security": [{ "BearerAuth": ["webhooks.write"] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["url", "events"], "properties": { "url": { "type": "string" }, "events": { "type": "array", "items": { "$ref": "#/components/schemas/WebhookEvent" } }, "secret": { "type": "string" }, "include_reseller_events": { "type": "boolean", "default": false, "description": "Owner-only opt-in for a node-wide endpoint to receive reseller customer events. Ordinary operator tokens cannot see or change opted-in endpoints." } } } } } },
        "responses": { "201": { "description": "Created (secret present when generated)" } }
      }
    },
    "/api/v1/webhooks/{id}": {
      "get": {
        "tags": ["webhooks"], "summary": "Get a webhook endpoint", "security": [{ "BearerAuth": ["webhooks.read"] }],
        "parameters": [{ "$ref": "#/components/parameters/WebhookID" }],
        "responses": { "200": { "description": "Endpoint", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebhookEndpoint" } } } } }
      },
      "patch": {
        "tags": ["webhooks"], "summary": "Update url/events/enabled or rotate the secret", "security": [{ "BearerAuth": ["webhooks.write"] }],
        "parameters": [{ "$ref": "#/components/parameters/WebhookID" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "url": { "type": "string" }, "events": { "type": "array", "items": { "$ref": "#/components/schemas/WebhookEvent" } }, "enabled": { "type": "boolean" }, "secret": { "type": "string", "description": "Empty string generates a new secret (returned once)" }, "include_reseller_events": { "type": "boolean", "description": "Owner-only change to node-wide reseller event fanout." } } } } } },
        "responses": { "200": { "description": "Updated" } }
      },
      "delete": {
        "tags": ["webhooks"], "summary": "Delete a webhook endpoint", "security": [{ "BearerAuth": ["webhooks.write"] }],
        "parameters": [{ "$ref": "#/components/parameters/WebhookID" }],
        "responses": { "200": { "description": "Deleted" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "description": "API token: `Authorization: Bearer wg_…` with scopes; separate from admin sessions." }
    },
    "parameters": {
...
      "WebhookID": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
    },
...
    "schemas": {
...
      "WebhookEvent": { "type": "string", "enum": ["user.created", "user.updated", "user.enabled", "user.disabled", "user.expired", "user.traffic_exceeded", "user.first_connected", "device.created", "device.deleted", "node.started"] },
...
      "WebhookEndpoint": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "reseller_id": { "type": "string", "description": "Present for a reseller-owned endpoint; absent for a node-wide endpoint." },
          "include_reseller_events": { "type": "boolean", "description": "Owner-approved subscription to reseller customer events for this node-wide endpoint. Always false for a reseller endpoint." },
          "url": { "type": "string" },
          "enabled": { "type": "boolean" },
          "events": { "type": "array", "items": { "$ref": "#/components/schemas/WebhookEvent" } },
          "created_at": { "type": "string", "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds." },
          "stats": {
            "type": "object",
            "properties": {
              "pending": { "type": "integer" },
              "delivered": { "type": "integer" },
              "dead": { "type": "integer" }
            }
          }
        }
      }
    }
  },
...
}